# Conflicts:
#	.idea/workspace.xml
#	cmake-build-debug/svm

+++ byteorder_helper.h
@@ -5,12 +5,12 @@
 #ifndef SVM_BYTE_HELPER_H
 #define SVM_BYTE_HELPER_H
 
-#include <cstdint>
+#include <stdint.h>
 #include <cstdlib>
 #include <cstdio>
 
 
-void flip(uint8_t* data, size_t size);
+void flip(uint8_t* data, unsigned int size);
 void fread_uint32_with_flip(uint32_t* i, FILE* f);
 
 #endif //SVM_BYTE_HELPER_H
